Markdown Tables

Returns a string of tables in markdown format ready for inclusion in a report.

md_tables(sub = get_sub(), main = get_main(), report = get_report(),
  headings = list(character(0)), drop = list(character(0)), nheaders = 0L,
  header1 = 3L, locale = "en", ask = getOption("subfoldr.ask", TRUE))

Arguments

sub
A string of the path to the subfolders to save the object (by default = "").

main
A string of the main subfolder.

report
A string indicating the report folder to copy the csv files.

headings
A list of named character vectors.

drop
A list of character vectors specify the subfolders to drop.

nheaders
An count of the number of headings to assign headers to.

header1
A count of the heading level for the first header.

locale
A string of the locale.

ask
A string indicating whether to ask before creating a sub directory.

Value

A string of the report templates in markdown format ready for inclusion in a report.

Details

The names in the character vectors in headings indicate the new headings for each subfolder. By default missing subfolders receive their current name with the first letter of each word capitalized. Subfolders with the name "" do not receive a heading. The order of the subfolders indicates the order in which they should appear. By default missing subfolders appear in alphabetical order. The first named character vector is applied to the highest level of subfolders starting at sub and so on. The number of character vectors indicates the number of levels that should receive headings. By default the highest level of subfolders are considered to be third order headings.

The elements in the character vectors in drop indicate the subfolders to exclude from the report. Again the number of the character vector indicates the level to which it applies.